refactor: optimize import
